examples/Kafka-TIP/requirements.txt
kafka-python3==3.0.0
mininet